# Use Blosc special values in empty arrays