--- a/docs/sphinx/development/index.rst
+++ b/docs/sphinx/development/index.rst
@@ -7,5 +7,3 @@ Development
     contributing
     roadmap
     code_of_conduct
-    release-notes
-