… CMakeLists.txt
… RELEASE_NOTES.rst
… bench/CMakeLists.txt
… bench/bench_get_slice.c
… caterva/caterva_blosc.h
… caterva/caterva_plainbuffer.h
… caterva/caterva_utils.c
… caterva/caterva_utils.h
… docs/doxygen/Doxyfile
… docs/requirements.txt
… docs/sphinx/_static/css/custom.css
… docs/sphinx/conf.py
… docs/sphinx/getting-started/installation.rst
… docs/sphinx/getting_started/index.rst
… docs/sphinx/getting_started/installation.rst
… docs/sphinx/getting_started/overview.png
… docs/sphinx/getting_started/overview.rst
… docs/sphinx/getting_started/tutorials/basics.rst
… docs/sphinx/getting_started/tutorials/index.rst
… docs/sphinx/index.rst
… docs/sphinx/reference/array.rst
… docs/sphinx/reference/context.rst
… docs/sphinx/reference/index.rst
… docs/sphinx/release_notes/index.rst
… tests/cutest.h
… tests/test_append.c
… tests/test_common.h
… tests/test_copy.c
… tests/test_full.c
… tests/test_get_slice.c