- [x] Use the Blosc machinery for special values to create empty arrays.
- [x] Allow the users to update array elements.