@aleixalcacer 2021-05-19
The results of the new benchmark file (bench/bench_get_slice.c) are:

- master
  ```
  from_buffer: 3.3 s
  get_slice: 0.79 s
  ```
- this PR
  ```
  from_buffer: 2.4 s
  get_slice: 0.59 s
  ```

As we can see, there exists a performance improvement!